+++ sgp4-predict/src/lib.rs
@@ -95,7 +95,7 @@ pub use sgp4::{Classification, Elements};
 
 pub use crate::{
     angle::{Degrees, Radians},
-    aoi::{AoiIter, AoiIterOpts, AoiWindow, Area, Error as AoiError, FillRule, Polygon},
+    aoi::{AoiIter, AoiIterOpts, AoiWindow, Area, Error as AoiError, FillRule, Polygon, Rectangle},
     apsides::{Apsis, ApsisEvent, ApsisIter, ApsisIterOpts},
     detect::Error as DetectError,
     frames::{EcefState, EnuState, Geodetic, LatLon, TemeState},